refactor: organize game modules, split out play from rules
