--- src/app/ai/constants.js
@@ -1,10 +1,0 @@
-export const AI_TYPES = {
-  LOWEST_CARD: "lowest_card",
-  HIGHEST_CARD: "highest_card",
-  LOWEST_VALUE: "lowest_value",
-  HIGHEST_VALUE: "highest_value",
-  RANDOM: "random",
-  PASS: "pass",
-  PRIORITIZED_COMBO: "prioritized_combo",
-  RANDOM_COMBO: "random_combo",
-};
