+++ src/app/ai/index.js
@@ -9,3 +9,14 @@ export { AI_PERSONAS } from "./personas.js";
 export { PrioritizedComboAI } from "./prioritized_combo.js";
 export { RandomAI } from "./random.js";
 export { RandomComboAI } from "./random_combo.js";
+
+export const AI_TYPES = {
+  LOWEST_CARD: "lowest_card",
+  HIGHEST_CARD: "highest_card",
+  LOWEST_VALUE: "lowest_value",
+  HIGHEST_VALUE: "highest_value",
+  RANDOM: "random",
+  PASS: "pass",
+  PRIORITIZED_COMBO: "prioritized_combo",
+  RANDOM_COMBO: "random_combo",
+};
