+++ src/app/ai/lowest_value.js
@@ -1,5 +1,5 @@
 import { AI } from "./base.js";
-import { AI_TYPES } from "./constants.js";
+import { AI_TYPES } from "./index.js";
 
 export class LowestValueAI extends AI {
   /**
