+++ src/app/ai/random_combo.js
@@ -1,6 +1,6 @@
 import { AI } from "./base.js";
 import { ComboAI } from "./combo.js";
-import { AI_TYPES } from "./constants.js";
+import { AI_TYPES } from "./index.js";
 
 export class RandomComboAI extends ComboAI {
   /**
