+++ src/app/game/game.js
@@ -1,7 +1,7 @@
-import { COMBINATION_TYPES, PLAYER_TYPES, RANKS } from "./constants.js";
 import { Card, Deck } from "./deck.js";
-import { createPlayer } from "./player/index.js";
-import { log } from "./utils.js";
+import { createPlayer, PLAYER_TYPES } from "../player/index.js";
+import { Rules } from "./rules.js";
+import { log } from "../utils.js";
 
 export class Game {
   STATE_KEY = "13gs";
@@ -42,6 +42,7 @@ export class Game {
       playerPersonas: [],
     };
     this.id = crypto.randomUUID();
+    this.rules = new Rules();
     this.stateKey = stateKey;
   }
 
@@ -78,22 +79,6 @@ export class Game {
   }
 
   /**
-   * Gets the AI player from the game's state.
-   * @returns {AIPlayer} The current AI player instance.
-   */
-  firstAIPlayer() {
-    return this.gameState.players.find((p) => p.type === PLAYER_TYPES.AI);
-  }
-
-  /**
-   * Gets the human player from the game's state.
-   * @returns {HumanPlayer} The current human player instance.
-   */
-  firstHumanPlayer() {
-    return this.gameState.players.find((p) => p.type === PLAYER_TYPES.HUMAN);
-  }
-
-  /**
    * Deals player hands for each player.
    */
   deal() {
@@ -123,18 +108,19 @@ export class Game {
   }
 
   /**
-   * Determines the combination type of a set of cards.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {string} The combination type.
+   * Gets the AI player from the game's state.
+   * @returns {AIPlayer} The current AI player instance.
    */
-  getCombinationType(cards) {
-    if (this.isSingle(cards)) return COMBINATION_TYPES.SINGLE;
-    if (this.isPair(cards)) return COMBINATION_TYPES.PAIR;
-    if (this.isTriple(cards)) return COMBINATION_TYPES.TRIPLE;
-    if (this.isStraight(cards)) return COMBINATION_TYPES.STRAIGHT;
-    if (this.isConsecutivePairs(cards)) return COMBINATION_TYPES.CONSECUTIVE_PAIRS;
-    if (this.isFourOfAKind(cards)) return COMBINATION_TYPES.FOUR_OF_A_KIND;
-    return COMBINATION_TYPES.INVALID;
+  firstAIPlayer() {
+    return this.gameState.players.find((p) => p.type === PLAYER_TYPES.AI);
+  }
+
+  /**
+   * Gets the human player from the game's state.
+   * @returns {HumanPlayer} The current human player instance.
+   */
+  firstHumanPlayer() {
+    return this.gameState.players.find((p) => p.type === PLAYER_TYPES.HUMAN);
   }
 
   /**
@@ -154,188 +140,6 @@ export class Game {
     this.gameState.selectedCards = [];
     this.deal();
     this.hooks.onGameInit(this);
-  }
-
-  /**
-   * Checks if a combination of cards is a bomb that can beat a pair of 2s.
-   * A bomb against a pair of 2s is four consecutive pairs.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the cards form a bomb that beats a pair of 2s.
-   */
-  isBombForPairOfTwos(cards) {
-    const combinationType = this.getCombinationType(cards);
-    if (combinationType === COMBINATION_TYPES.CONSECUTIVE_PAIRS && cards.length === 8) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Checks if a combination of cards is a bomb that can beat a single 2.
-   * A bomb against a single 2 is either a four-of-a-kind or three consecutive pairs.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the cards form a bomb that beats a single 2.
-   */
-  isBombForSingleTwo(cards) {
-    const combinationType = this.getCombinationType(cards);
-    if (combinationType === COMBINATION_TYPES.FOUR_OF_A_KIND) {
-      return true;
-    }
-    if (combinationType === COMBINATION_TYPES.CONSECUTIVE_PAIRS && cards.length === 6) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Checks if a combination is consecutive pairs.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is consecutive pairs.
-   */
-  isConsecutivePairs(cards) {
-    if (cards.length < 6 || cards.length % 2 !== 0 || cards.some((card) => card.rank === "2")) {
-      return false;
-    }
-    const pairs = [];
-    for (let i = 0; i < cards.length; i += 2) {
-      const pair = [cards[i], cards[i + 1]];
-      if (!this.isPair(pair)) {
-        return false;
-      }
-      pairs.push(pair);
-    }
-    for (let i = 0; i < pairs.length - 1; i++) {
-      const rankIndex1 = RANKS.indexOf(pairs[i][0].rank);
-      const rankIndex2 = RANKS.indexOf(pairs[i + 1][0].rank);
-      if (rankIndex2 !== rankIndex1 + 1) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  /**
-   * Checks if a combination is a four-of-a-kind.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is a four-of-a-kind.
-   */
-  isFourOfAKind(cards) {
-    return cards.length === 4 && Card.allSameRank(cards);
-  }
-
-  /**
-   * Checks if a combination is a pair.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is a pair.
-   */
-  isPair(cards) {
-    return cards.length === 2 && Card.allSameRank(cards);
-  }
-
-  /**
-   * Checks if a combination is a single card.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is a single.
-   */
-  isSingle(cards) {
-    return cards.length === 1;
-  }
-
-  /**
-   * Checks if a combination is a straight.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is a straight.
-   */
-  isStraight(cards) {
-    if (cards.length < 3) {
-      return false;
-    }
-    if (cards.some((card) => card.rank === "2")) {
-      return false;
-    }
-    for (let i = 0; i < cards.length - 1; i++) {
-      const rankIndex1 = RANKS.indexOf(cards[i].rank);
-      const rankIndex2 = RANKS.indexOf(cards[i + 1].rank);
-      if (rankIndex2 !== rankIndex1 + 1) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  /**
-   * Checks if a combination is a triple.
-   * @param {Array<Card>} cards The cards to check.
-   * @returns {boolean} True if the combination is a triple.
-   */
-  isTriple(cards) {
-    return cards.length === 3 && Card.allSameRank(cards);
-  }
-
-  /**
-   * Checks if a play is valid according to the game rules.
-   * @param {Array<Card>} selectedCards The cards the player has selected.
-   * @param {Array<Card>} playPile The cards currently in the play pile.
-   * @param {Array<Card>} playerHand The hand of the player making the play.
-   * @returns {boolean} True if the play is valid.
-   */
-  isValidPlay(selectedCards, playPile, playerHand, currentTurn, allPlayerHands) {
-    // Check if all selected cards are in the current player's hand
-    for (const selectedCard of selectedCards) {
-      if (!playerHand.some((card) => card.rank === selectedCard.rank && card.suit === selectedCard.suit)) {
-        return false;
-      }
-    }
-
-    const playPileCombinationType = this.getCombinationType(playPile);
-    const selectedCombinationType = this.getCombinationType(selectedCards);
-
-    if (selectedCombinationType === COMBINATION_TYPES.INVALID) {
-      return false;
-    }
-
-    // Bomb logic: A bomb can only be played on a 2 or a pair of 2s.
-    if (
-      selectedCombinationType === COMBINATION_TYPES.FOUR_OF_A_KIND ||
-      selectedCombinationType === COMBINATION_TYPES.CONSECUTIVE_PAIRS
-    ) {
-      if (playPile.length === 0) return false; // Bombs cannot be played on an empty pile
-
-      const topCard = playPile[playPile.length - 1];
-      if (topCard.rank !== "2") return false; // Bombs can only be played on 2s
-
-      if (playPile.length === 1) {
-        return this.isBombForSingleTwo(selectedCards);
-      }
-
-      if (playPile.length === 2) {
-        return this.isBombForPairOfTwos(selectedCards);
-      }
-
-      return false; // Bombs cannot be played on other combinations
-    }
-
-    // Standard play validation
-    if (playPile.length === 0) {
-      if (currentTurn === 0) {
-        const lowestCard = Card.findLowest(allPlayerHands);
-        return selectedCards.some((card) => card.value === lowestCard.value);
-      }
-      return true;
-    }
-
-    if (selectedCombinationType !== playPileCombinationType) {
-      return false;
-    }
-
-    if (selectedCards.length !== playPile.length) {
-      return false;
-    }
-
-    const highestSelectedCard = selectedCards.reduce((max, card) => (card.value > max.value ? card : max));
-    const highestPlayPileCard = playPile.reduce((max, card) => (card.value > max.value ? card : max));
-
-    return highestSelectedCard.value > highestPlayPileCard.value;
   }
 
   /**
@@ -439,7 +243,7 @@ export class Game {
     if (parsedState && ui) {
       // Store playerTypes from the parsed state before overwriting this.gameState
       const loadedPlayerTypes = parsedState.gameState.players.map((p) => p.type);
-      const loadedPlayerPersonas = parsedState.gameState.players.map((p) => (p.ai ? p.ai.type : null));
+      const loadedPlayerPersonas = parsedState.gameState.players.map((p) => (p.ai ? p.ai.persona : null));
 
       this.id = parsedState.id;
       this.stateKey = parsedState.stateKey;
