+++ src/app/game/index.js
@@ -1,3 +1,7 @@
+export { Card, Deck } from "./deck.js";
+export { Game } from "./game.js";
+export { Rules } from "./rules.js";
+
 export const SUITS = ["♠", "♣", "♦", "♥"];
 export const RANKS = ["3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K", "A", "2"];
 export const COMBINATION_TYPES = {
@@ -8,8 +12,4 @@ export const COMBINATION_TYPES = {
   CONSECUTIVE_PAIRS: "consecutive_pairs",
   FOUR_OF_A_KIND: "four_of_a_kind",
   INVALID: "invalid",
-};
-export const PLAYER_TYPES = {
-  AI: "ai",
-  HUMAN: "human",
 };
