+++ src/app/player/index.js
@@ -2,3 +2,8 @@ export { AIPlayer } from "./ai.js";
 export { Player } from "./base.js";
 export { createPlayer } from "./factory.js";
 export { HumanPlayer } from "./human.js";
+
+export const PLAYER_TYPES = {
+  AI: "ai",
+  HUMAN: "human",
+};
