+++ tests/ai/test_base.js
@@ -1,7 +1,5 @@
 import { AI } from "../../src/app/ai/index.js";
-import { COMBINATION_TYPES } from "../../src/app/constants.js";
-import { Card } from "../../src/app/deck.js";
-import { Game } from "../../src/app/game.js";
+import { Card, Game, COMBINATION_TYPES } from "../../src/app/game/index.js";
 
 describe("AI", () => {
   let game, ai;
