+++ tests/ai/test_lowest_card.js
@@ -1,6 +1,5 @@
 import { LowestCardAI } from "../../src/app/ai/index.js";
-import { Card } from "../../src/app/deck.js";
-import { Game } from "../../src/app/game.js";
+import { Card, Game } from "../../src/app/game/index.js";
 
 describe("LowestCardAI", () => {
   let game, lowestCardAI;
