+++ tests/ai/test_random_combo.js
@@ -1,6 +1,5 @@
 import { RandomComboAI } from "../../src/app/ai/index.js";
-import { Card } from "../../src/app/deck.js";
-import { Game } from "../../src/app/game.js";
+import { Card, Game } from "../../src/app/game/index.js";
 
 describe("RandomComboAI", () => {
   let game, randomComboAI, strategy1, strategy2;
