+++ tests/game/test_game.js
@@ -1,8 +1,6 @@
-import { Card } from "../src/app/deck.js";
-import { Game } from "../src/app/game.js";
-import { MockDeck, MockAI, MockUI, MockPlayer, MockLocalStorage } from "./mocks.js";
-import { HumanPlayer, AIPlayer } from "../src/app/player/index.js";
-import { COMBINATION_TYPES, PLAYER_TYPES } from "../src/app/constants.js";
+import { Card, Game } from "../../src/app/game/index.js";
+import { HumanPlayer, AIPlayer, PLAYER_TYPES } from "../../src/app/player/index.js";
+import { MockDeck, MockUI, MockPlayer } from "../mocks.js";
 
 describe("Game", () => {
   let game;
@@ -134,115 +132,6 @@ describe("Game", () => {
       expect(game.id).to.not.equal(originalGameId);
       expect(game.gameState.gamesWon).to.deep.equal([0, 0]);
       expect(game.gameState.roundNumber).to.equal(1); // Check that init() was called
-    });
-  });
-
-  describe("Card Combination Logic", () => {
-    it("getCombinationType() should identify all valid combination types", () => {
-      expect(game.getCombinationType([new Card("A", "♠")])).to.equal(COMBINATION_TYPES.SINGLE);
-      expect(game.getCombinationType([new Card("A", "♠"), new Card("A", "♦")])).to.equal(COMBINATION_TYPES.PAIR);
-      expect(game.getCombinationType([new Card("A", "♠"), new Card("A", "♣"), new Card("A", "♦")])).to.equal(
-        COMBINATION_TYPES.TRIPLE,
-      );
-      expect(game.getCombinationType([new Card("A", "♠"), new Card("A", "♣"), new Card("A", "♦"), new Card("A", "♥")])).to.equal(
-        COMBINATION_TYPES.FOUR_OF_A_KIND,
-      );
-      expect(game.getCombinationType([new Card("3", "♠"), new Card("4", "♦"), new Card("5", "♣")])).to.equal(
-        COMBINATION_TYPES.STRAIGHT,
-      );
-      const consecutivePairs = [
-        new Card("3", "♠"),
-        new Card("3", "♦"),
-        new Card("4", "♣"),
-        new Card("4", "♥"),
-        new Card("5", "♠"),
-        new Card("5", "♦"),
-      ];
-      expect(game.getCombinationType(consecutivePairs)).to.equal(COMBINATION_TYPES.CONSECUTIVE_PAIRS);
-      expect(game.getCombinationType([new Card("K", "♠"), new Card("A", "♠")])).to.equal(COMBINATION_TYPES.INVALID);
-    });
-
-    it("isSingle() should correctly identify a single", () => {
-      expect(game.isSingle([new Card("A", "♠")])).to.be.true;
-      expect(game.isSingle([new Card("A", "♠"), new Card("K", "♠")])).to.be.false;
-    });
-
-    it("isPair() should correctly identify a pair", () => {
-      expect(game.isPair([new Card("A", "♠"), new Card("A", "♦")])).to.be.true;
-      expect(game.isPair([new Card("A", "♠"), new Card("K", "♠")])).to.be.false;
-    });
-
-    it("isTriple() should correctly identify a triple", () => {
-      expect(game.isTriple([new Card("A", "♠"), new Card("A", "♦"), new Card("A", "♣")])).to.be.true;
-      expect(game.isTriple([new Card("A", "♠"), new Card("K", "♠"), new Card("Q", "♠")])).to.be.false;
-    });
-
-    it("isFourOfAKind() should correctly identify a four-of-a-kind", () => {
-      expect(game.isFourOfAKind([new Card("A", "♠"), new Card("A", "♦"), new Card("A", "♣"), new Card("A", "♥")])).to.be.true;
-      expect(game.isFourOfAKind([new Card("A", "♠"), new Card("K", "♠"), new Card("Q", "♠"), new Card("J", "♠")])).to.be.false;
-    });
-
-    it("isStraight() should correctly identify straights", () => {
-      expect(game.isStraight([new Card("3", "♠"), new Card("4", "♦"), new Card("5", "♣")])).to.be.true;
-      expect(game.isStraight([new Card("3", "♠"), new Card("5", "♦"), new Card("6", "♣")])).to.be.false;
-      expect(game.isStraight([new Card("K", "♠"), new Card("A", "♦"), new Card("2", "♣")])).to.be.false;
-    });
-
-    it("isConsecutivePairs() should correctly identify consecutive pairs", () => {
-      const validPairs = [
-        new Card("3", "♠"),
-        new Card("3", "♦"),
-        new Card("4", "♣"),
-        new Card("4", "♥"),
-        new Card("5", "♠"),
-        new Card("5", "♦"),
-      ];
-      const invalidPairs = [
-        new Card("3", "♠"),
-        new Card("3", "♦"),
-        new Card("4", "♣"),
-        new Card("5", "♥"),
-        new Card("5", "♠"),
-        new Card("5", "♦"),
-      ];
-      const nonConsecutive = [new Card("3", "♠"), new Card("3", "♦"), new Card("5", "♣"), new Card("5", "♥")];
-      const withTwos = [new Card("A", "♠"), new Card("A", "♦"), new Card("2", "♣"), new Card("2", "♥")];
-      expect(game.isConsecutivePairs(validPairs)).to.be.true;
-      expect(game.isConsecutivePairs(invalidPairs)).to.be.false;
-      expect(game.isConsecutivePairs(nonConsecutive)).to.be.false;
-      expect(game.isConsecutivePairs(withTwos)).to.be.false;
-    });
-
-    it("isBombForSingleTwo() should identify correct bombs for a single two", () => {
-      const fourOfAKind = [new Card("A", "♠"), new Card("A", "♦"), new Card("A", "♣"), new Card("A", "♥")];
-      const threeConsecutivePairs = [
-        new Card("3", "♠"),
-        new Card("3", "♦"),
-        new Card("4", "♣"),
-        new Card("4", "♥"),
-        new Card("5", "♠"),
-        new Card("5", "♦"),
-      ];
-      const invalidBomb = [new Card("A", "♠"), new Card("K", "♠"), new Card("Q", "♠")];
-      expect(game.isBombForSingleTwo(fourOfAKind)).to.be.true;
-      expect(game.isBombForSingleTwo(threeConsecutivePairs)).to.be.true;
-      expect(game.isBombForSingleTwo(invalidBomb)).to.be.false;
-    });
-
-    it("isBombForPairOfTwos() should identify correct bombs for a pair of twos", () => {
-      const fourConsecutivePairs = [
-        new Card("3", "♠"),
-        new Card("3", "♦"),
-        new Card("4", "♣"),
-        new Card("4", "♥"),
-        new Card("5", "♠"),
-        new Card("5", "♦"),
-        new Card("6", "♣"),
-        new Card("6", "♦"),
-      ];
-      const invalidBomb = [new Card("A", "♠"), new Card("A", "♦"), new Card("A", "♣"), new Card("A", "♥")];
-      expect(game.isBombForPairOfTwos(fourConsecutivePairs)).to.be.true;
-      expect(game.isBombForPairOfTwos(invalidBomb)).to.be.false;
     });
   });
 
@@ -404,117 +293,6 @@ describe("Game", () => {
       expect(newGame.gameState.playPile[0].suit).to.equal("♠");
       expect(newGame.gameState.selectedCards[0].rank).to.equal("A");
       expect(newGame.gameState.selectedCards[0].suit).to.equal("♠");
-    });
-  });
-
-  describe("Play Validation", () => {
-    it("should not allow playing cards that are not in the player's hand", () => {
-      const playerHand = [new Card("3", "♠")];
-      const selectedCards = [new Card("4", "♦")];
-      const isValid = game.isValidPlay(selectedCards, [], playerHand, 1, [playerHand]);
-      expect(isValid).to.be.false;
-    });
-
-    it("should require the lowest card to be played on the first turn", () => {
-      const allPlayerHands = [[new Card("4", "♠")], [new Card("3", "♦")]];
-      const playerHand = allPlayerHands[0];
-      const selectedCards = [new Card("4", "♠")];
-      const isValid = game.isValidPlay(selectedCards, [], playerHand, 0, allPlayerHands);
-      expect(isValid).to.be.false;
-    });
-
-    it("should allow any valid combination on an empty pile (after first turn)", () => {
-      const playerHand = [new Card("5", "♠")];
-      const isValid = game.isValidPlay([new Card("5", "♠")], [], playerHand, 1, [playerHand]);
-      expect(isValid).to.be.true;
-    });
-
-    it("should require the same combination type as the play pile", () => {
-      const playerHand = [new Card("5", "♠"), new Card("5", "♦")];
-      const playPile = [new Card("4", "♠")];
-      const isValid = game.isValidPlay(playerHand, playPile, playerHand, 1, [playerHand]);
-      expect(isValid).to.be.false;
-    });
-
-    it("should require a higher-value card for singles", () => {
-      const playerHand = [new Card("5", "♠")];
-      const playPile = [new Card("4", "♠")];
-      const isValid = game.isValidPlay(playerHand, playPile, playerHand, 1, [playerHand]);
-      expect(isValid).to.be.true;
-    });
-
-    it("should require a higher-value pair", () => {
-      const playerHand = [new Card("5", "♠"), new Card("5", "♦")];
-      const playPile = [new Card("4", "♠"), new Card("4", "♦")];
-      const isValid = game.isValidPlay(playerHand, playPile, playerHand, 1, [playerHand]);
-      expect(isValid).to.be.true;
-    });
-
-    it("should require a higher-value triple", () => {
-      const playerHand = [new Card("5", "♠"), new Card("5", "♦"), new Card("5", "♣")];
-      const playPile = [new Card("4", "♠"), new Card("4", "♦"), new Card("4", "♣")];
-      const isValid = game.isValidPlay(playerHand, playPile, playerHand, 1, [playerHand]);
-      expect(isValid).to.be.true;
-    });
-
-    it("should require a higher-value straight", () => {
-      const playerHand = [new Card("4", "♠"), new Card("5", "♦"), new Card("6", "♣")];
-      const playPile = [new Card("3", "♠"), new Card("4", "♦"), new Card("5", "♣")];
-      const isValid = game.isValidPlay(playerHand, playPile, playerHand, 1, [playerHand]);
-      expect(isValid).to.be.true;
-    });
-
-    describe("Bomb Logic", () => {
-      it("should allow a four-of-a-kind to beat a single 2", () => {
-        const bomb = [new Card("4", "♠"), new Card("4", "♦"), new Card("4", "♣"), new Card("4", "♥")];
-        const playPile = [new Card("2", "♠")];
-        const isValid = game.isValidPlay(bomb, playPile, bomb, 1, [bomb]);
-        expect(isValid).to.be.true;
-      });
-
-      it("should allow three consecutive pairs to beat a single 2", () => {
-        const bomb = [
-          new Card("3", "♠"),
-          new Card("3", "♦"),
-          new Card("4", "♣"),
-          new Card("4", "♥"),
-          new Card("5", "♠"),
-          new Card("5", "♦"),
-        ];
-        const playPile = [new Card("2", "♠")];
-        const isValid = game.isValidPlay(bomb, playPile, bomb, 1, [bomb]);
-        expect(isValid).to.be.true;
-      });
-
-      it("should allow four consecutive pairs to beat a pair of 2s", () => {
-        const bomb = [
-          new Card("3", "♠"),
-          new Card("3", "♦"),
-          new Card("4", "♣"),
-          new Card("4", "♥"),
-          new Card("5", "♠"),
-          new Card("5", "♦"),
-          new Card("6", "♣"),
-          new Card("6", "♦"),
-        ];
-        const playPile = [new Card("2", "♠"), new Card("2", "♦")];
-        const isValid = game.isValidPlay(bomb, playPile, bomb, 1, [bomb]);
-        expect(isValid).to.be.true;
-      });
-
-      it("should not allow a bomb on a non-2 card", () => {
-        const bomb = [new Card("4", "♠"), new Card("4", "♦"), new Card("4", "♣"), new Card("4", "♥")];
-        const playPile = [new Card("K", "♠")];
-        const isValid = game.isValidPlay(bomb, playPile, bomb, 1, [bomb]);
-        expect(isValid).to.be.false;
-      });
-
-      it("should not allow a four-of-a-kind to bomb a pair of 2s", () => {
-        const bomb = [new Card("4", "♠"), new Card("4", "♦"), new Card("4", "♣"), new Card("4", "♥")];
-        const playPile = [new Card("2", "♠"), new Card("2", "♦")];
-        const isValid = game.isValidPlay(bomb, playPile, bomb, 1, [bomb]);
-        expect(isValid).to.be.false;
-      });
     });
   });
 });
