+++ tests/index.html
@@ -32,8 +32,9 @@
         },
       };
     </script>
-    <script type="module" src="./test_deck.js"></script>
-    <script type="module" src="./test_game.js"></script>
+    <script type="module" src="./game/test_deck.js"></script>
+    <script type="module" src="./game/test_game.js"></script>
+    <script type="module" src="./game/test_rules.js"></script>
     <script type="module" src="./player/test_base.js"></script>
     <script type="module" src="./player/test_ai.js"></script>
     <script type="module" src="./player/test_human.js"></script>
