+++ tests/player/test_ai.js
@@ -1,6 +1,5 @@
-import { PLAYER_TYPES } from "../../src/app/constants.js";
-import { Card } from "../../src/app/deck.js";
-import { AIPlayer } from "../../src/app/player/index.js";
+import { Card } from "../../src/app/game/index.js";
+import { AIPlayer, PLAYER_TYPES } from "../../src/app/player/index.js";
 import { MockAI, MockGame } from "../mocks.js";
 
 describe("AIPlayer", () => {
